COMPONENTE DE DIALOG, COMPLEXO EIN !

+++ src/app/heroes/components/heroes/heroes.component.ts
@@ -1,3 +1,5 @@
+import { ConfimationDialogComponent } from './../../../core/components/confimation-dialog/confimation-dialog.component';
+import { DialogData } from './../../../core/models/dialog-data.model';
 
 import { HeroService } from '../../../core/services/hero.service';
 /* eslint-disable @typescript-eslint/no-empty-function */
@@ -5,6 +7,8 @@ import { Hero } from '../../../core/models/hero.model';
 import { Component, OnInit } from '@angular/core';
 
 import { HEROES } from '../../../core/services/mock-heroes';
+import { MatDialog } from '@angular/material/dialog';
+
 
 @Component({
   selector: 'app-heroes',
@@ -16,7 +20,9 @@ export class HeroesComponent implements OnInit {
   heroes: Hero[] = [];
   displayedColumns: string[] = ['id','name', 'actions']
 
-  constructor (private heroService : HeroService,
+  constructor (
+    private heroService : HeroService,
+    private dialog : MatDialog
    ) { }
 
   ngOnInit(): void {
@@ -39,10 +45,19 @@ export class HeroesComponent implements OnInit {
   } */
 
   deleteHero( hero : Hero): void {
+    const dialog : DialogData = { cancelText : 'cancelar', confirmText :"sim",  content : `Deletar ${hero.name}`}
+
+    const dialogRef = this.dialog.open(ConfimationDialogComponent, {data : dialog , width : '600px'})
     //subscrevendo no observable e tratando o retorno
-   this.heroService.deleteHero(hero).subscribe( ()=> {
-    this.getHeroes()
-   });
+
+    dialogRef.afterClosed().subscribe( result => {
+      if (result) {
+        this.heroService.deleteHero(hero).subscribe( ()=> {
+          this.getHeroes()
+      })}}
+    )
+
+
 
   }
